Menambahkan keterangan pada notebook dan fixing main.py
